Fixed response types of billing endpoints

+++ lib/Model/ApiHourBillingCalculationPosition.php
@@ -69,7 +69,7 @@ class ApiHourBillingCalculationPosition implements ModelInterface, ArrayAccess, 
         'is_external_service' => 'bool',
         'is_service_needed' => 'bool',
         'display_path' => 'string',
-        'quantity' => 'int',
+        'quantity' => 'float',
         'sale_price' => 'float',
         'purchase_price' => 'float',
         'total_calculation' => 'float',
@@ -894,7 +894,7 @@ class ApiHourBillingCalculationPosition implements ModelInterface, ArrayAccess, 
     /**
      * Gets quantity
      *
-     * @return int|null
+     * @return float|null
      */
     public function getQuantity()
     {
@@ -904,7 +904,7 @@ class ApiHourBillingCalculationPosition implements ModelInterface, ArrayAccess, 
     /**
      * Sets quantity
      *
-     * @param int|null $quantity quantity
+     * @param float|null $quantity quantity
      *
      * @return self
      */
